backend/src/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
sqlparse==0.5.0
uuid==1.30
drf-yasg==1.21.7 